==> R_scripts/0.2_common_variables.R <==
#Set the parent directory (this should be the same directory as is set in the common_variables.sh script
parent_dir <- 'N:/JETTLES/VU_CNOT3_KD'

#set sample names

Total_sample_names <- c("CONTROL1", "CONTROL2", "CONTROL3", "SH331", "SH332", "SH333", "SH371", "SH372", "SH373")

#set read lengths used for library QC
#lengths <- 25:35
